Added high score functionality on menu screen and commented code
Added the class Global as a way to access data across states.
GameStateManager now holds a HashMap of Globals so that any state has
access to all the Globals in the HashMap. This allowed for
implementation of a high score global variable, which is displayed in
the MenuState. Also commented most of the code (have not comment HUD).